Added additional tests, created seperate file for MultiSelectValue
