+++ src/js/components/MultiSelect/__tests__/MultiSelect-test.tsx
@@ -1,15 +1,19 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import { axe } from 'jest-axe';
 import 'jest-axe/extend-expect';
 import 'regenerator-runtime/runtime';
 import 'jest-styled-components';
 import '@testing-library/jest-dom';
+import { createPortal, expectPortal } from '../../../utils/portal';
 
 import { Grommet } from '../..';
 import { MultiSelect } from '..';
 
-describe('MultiSelect Controlled', () => {
+describe('MultiSelect', () => {
+  window.scrollTo = jest.fn();
+  beforeEach(createPortal);
   test('should not have accessibility violations', async () => {
     const { container } = render(
       <Grommet>
@@ -32,5 +36,115 @@ describe('MultiSelect Controlled', () => {
     });
     expect(container.firstChild).toMatchSnapshot();
     expect(results).toHaveNoViolations();
+  });
+
+  test('defaultValue', () => {
+    const { container } = render(
+      <Grommet>
+        <MultiSelect options={['one', 'two']} defaultValue={['one']} />,
+      </Grommet>,
+    );
+
+    expect(container.firstChild).toMatchSnapshot();
+  });
+
+  test('children', () => {
+    const { container } = render(
+      <Grommet>
+        <MultiSelect options={[{ test: 'one' }, { test: 'two' }]}>
+          {(option) => <span>{option.test}</span>}
+        </MultiSelect>
+      </Grommet>,
+    );
+
+    expect(container.firstChild).toMatchSnapshot();
+  });
+
+  test('placeholder', () => {
+    const { container } = render(
+      <Grommet>
+        <MultiSelect
+          options={[{ test: 'one' }, { test: 'two' }]}
+          placeholder="placeholder text"
+        />
+      </Grommet>,
+    );
+
+    expect(container.firstChild).toMatchSnapshot();
+  });
+
+  test('disabled', async () => {
+    const { container } = render(
+      <Grommet>
+        <MultiSelect options={[1, 2]} disabled />
+      </Grommet>,
+    );
+
+    expect(container.firstChild).toMatchSnapshot();
+  });
+
+  test('disabled option', async () => {
+    const user = userEvent.setup();
+    render(
+      <Grommet>
+        <MultiSelect
+          id="test-select__drop"
+          options={[0, 1, 2]}
+          disabled={[1]}
+        />
+      </Grommet>,
+    );
+    // open MultiSelect
+    await user.click(screen.getByRole('button', { name: /Open Drop/i }));
+    // try to click all the options
+    await user.click(screen.getByRole('option', { name: /0/i }));
+    await user.click(screen.getByRole('option', { name: /1/i }));
+    await user.click(screen.getByRole('option', { name: /2/i }));
+
+    // only 2 options should be selected (0 and 2)
+    expectPortal('test-select__drop').toMatchSnapshot();
+  });
+
+  test('limit', async () => {
+    window.HTMLElement.prototype.scrollIntoView = jest.fn();
+    const user = userEvent.setup();
+    render(
+      <Grommet>
+        <MultiSelect id="test-select__drop" options={[0, 1, 2]} limit={2} />
+      </Grommet>,
+    );
+    // open MultiSelect
+    await user.click(screen.getByRole('button', { name: /Open Drop/i }));
+    // select 2 options
+    await user.click(screen.getByRole('option', { name: /0/i }));
+    await user.click(screen.getByRole('option', { name: /1/i }));
+    await user.click(screen.getByRole('option', { name: /2/i }));
+
+    // option 2 should be disabled
+    expectPortal('test-select__drop').toMatchSnapshot();
+  });
+
+  test('visibleSelection', async () => {
+    // Mock scrollIntoView since JSDOM doesn't do layout.
+    // https://github.com/jsdom/jsdom/issues/1695#issuecomment-449931788
+    window.HTMLElement.prototype.scrollIntoView = jest.fn();
+    const user = userEvent.setup();
+    const { container } = render(
+      <Grommet>
+        <MultiSelect options={[0, 1, 2]} visibleSelection />
+      </Grommet>,
+    );
+    // open MultiSelect
+    await user.click(screen.getByRole('button', { name: /Open Drop/i }));
+    // click all the options
+    await user.click(screen.getByRole('option', { name: /0/i }));
+    await user.click(screen.getByRole('option', { name: /1/i }));
+    await user.click(screen.getByRole('option', { name: /2/i }));
+
+    // close MultiSelect
+    await user.click(screen.getByRole('button', { name: /Close Select/i }));
+
+    // all options should be visible when drop is closed
+    expect(container.firstChild).toMatchSnapshot();
   });
 });
